# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Scheme Postings API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  displayName: string;
}

export interface User {
  id: string;
  displayName: string;
  name: string;
  tags: string[];
  createdAt: string;
  isAdmin?: boolean;
}

export type PostStatus = typeof PostStatus[keyof typeof PostStatus];


export const PostStatus = {
  draft: 'draft',
  'in-review': 'in-review',
  publish: 'publish',
} as const;

export interface Post {
  id: string;
  title: string;
  body: string;
  excerpt: string;
  authorId: string;
  authorName: string;
  status: PostStatus;
  category: string;
  tags: string[];
  createdAt: string;
  updatedAt: string;
  annotationCount: number;
  noteCount: number;
  readyForGallery?: boolean;
}

export interface PostInput {
  title?: string;
  body?: string;
  excerpt?: string;
  category?: string;
  tags?: string[];
}

export interface PostUpdate {
  title?: string;
  body?: string;
  excerpt?: string;
  category?: string;
  tags?: string[];
  status?: string;
}

export type StrokePointsItem = {
  x?: number;
  y?: number;
};

export interface Stroke {
  points?: StrokePointsItem[];
}

export interface TextAnnotation {
  x?: number;
  y?: number;
  text?: string;
}

export type AnnotationType = typeof AnnotationType[keyof typeof AnnotationType];


export const AnnotationType = {
  drawing: 'drawing',
  text: 'text',
  note: 'note',
} as const;

export interface Annotation {
  id: string;
  postId: string;
  type: AnnotationType;
  /** @nullable */
  text?: string | null;
  /** @nullable */
  noteType?: string | null;
  strokes?: Stroke[];
  texts?: TextAnnotation[];
  /** @nullable */
  imageData?: string | null;
  authorId: string;
  authorName: string;
  createdAt: string;
}

export type AnnotationInputType = typeof AnnotationInputType[keyof typeof AnnotationInputType];


export const AnnotationInputType = {
  drawing: 'drawing',
  text: 'text',
  note: 'note',
} as const;

export interface AnnotationInput {
  type: AnnotationInputType;
  text?: string;
  noteType?: string;
  strokes?: Stroke[];
  texts?: TextAnnotation[];
  imageData?: string;
}

export interface NoteInput {
  text: string;
  noteType?: string;
}

export interface Comment {
  id: string;
  postId: string;
  text: string;
  authorId: string;
  authorName: string;
  createdAt: string;
}

export interface CommentInput {
  text: string;
}

export type PinCardType = typeof PinCardType[keyof typeof PinCardType];


export const PinCardType = {
  post: 'post',
  annotationNode: 'annotationNode',
} as const;

export interface PinCard {
  id: string;
  type: PinCardType;
  /** @nullable */
  postId?: string | null;
  /** @nullable */
  parentCardId?: string | null;
  /** @nullable */
  annotationId?: string | null;
  title?: string;
  /** @nullable */
  authorName?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  category?: string | null;
  tags?: string[];
  /** @nullable */
  noteText?: string | null;
  /** @nullable */
  annotationType?: string | null;
  /** @nullable */
  annotationPreview?: string | null;
  x?: number;
  y?: number;
  /** @nullable */
  groupId?: string | null;
}

export interface PinConnection {
  from: string;
  to: string;
  system?: boolean;
}

export interface PinboardData {
  workspace: string;
  cards: PinCard[];
  connections: PinConnection[];
  updatedAt: string;
}

export interface PinboardInput {
  cards?: PinCard[];
  connections?: PinConnection[];
}

export type ListPostsParams = {
authorId?: string;
status?: string;
};