# lib/api-zod/src/generated/types/annotation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Scheme Postings API
 * OpenAPI spec version: 0.1.0
 */
import type { AnnotationType } from './annotationType';
import type { Stroke } from './stroke';
import type { TextAnnotation } from './textAnnotation';

export interface Annotation {
  id: string;
  postId: string;
  type: AnnotationType;
  /** @nullable */
  text?: string | null;
  /** @nullable */
  noteType?: string | null;
  strokes?: Stroke[];
  texts?: TextAnnotation[];
  /** @nullable */
  imageData?: string | null;
  authorId: string;
  authorName: string;
  createdAt: string;
}