# lib/api-zod/src/generated/types/annotationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Scheme Postings API
 * OpenAPI spec version: 0.1.0
 */
import type { AnnotationInputType } from './annotationInputType';
import type { Stroke } from './stroke';
import type { TextAnnotation } from './textAnnotation';

export interface AnnotationInput {
  type: AnnotationInputType;
  text?: string;
  noteType?: string;
  strokes?: Stroke[];
  texts?: TextAnnotation[];
  imageData?: string;
}