# lib/api-zod/src/generated/types/pinCard.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Scheme Postings API
 * OpenAPI spec version: 0.1.0
 */
import type { PinCardType } from './pinCardType';

export interface PinCard {
  id: string;
  type: PinCardType;
  /** @nullable */
  postId?: string | null;
  /** @nullable */
  parentCardId?: string | null;
  /** @nullable */
  annotationId?: string | null;
  title?: string;
  /** @nullable */
  authorName?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  category?: string | null;
  tags?: string[];
  /** @nullable */
  noteText?: string | null;
  /** @nullable */
  annotationType?: string | null;
  /** @nullable */
  annotationPreview?: string | null;
  x?: number;
  y?: number;
  /** @nullable */
  groupId?: string | null;
}