# lib/api-zod/src/generated/types/pinConnection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Scheme Postings API
 * OpenAPI spec version: 0.1.0
 */

export interface PinConnection {
  from: string;
  to: string;
  system?: boolean;
}