# lib/api-zod/src/generated/types/post.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Scheme Postings API
 * OpenAPI spec version: 0.1.0
 */
import type { PostStatus } from './postStatus';

export interface Post {
  id: string;
  title: string;
  body: string;
  excerpt: string;
  authorId: string;
  authorName: string;
  status: PostStatus;
  category: string;
  tags: string[];
  createdAt: string;
  updatedAt: string;
  annotationCount: number;
  noteCount: number;
  readyForGallery?: boolean;
}