# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Scheme Postings API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  displayName: string;
  name: string;
  tags: string[];
  createdAt: string;
  isAdmin?: boolean;
}